requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pyaudio>=0.2.13
mss>=9.0.0
